+++ query/src/Lexer.h
@@ -13,6 +13,8 @@
 
 #include <string>
 #include <utility>
+#include <vector>
+#include <unordered_map>
 
 class Lexer {
 public:
@@ -38,6 +40,12 @@ public:
      */
     [[nodiscard]] auto EndOfInput() -> bool;
 
+    /** 
+     * Goes through the query and returns a frequency count of all the tokens
+    */
+    [[nodiscard]] auto GetTokenFrequencies() const -> std::unordered_map<std::string, int>;
+
+
 private:
     std::string input_;                                   // The input string to be tokenized
     size_t position_ = 0;                                 // Current position in the input string
@@ -60,6 +68,7 @@ private:
     auto LexWordOrKeyword() -> Token;  // Lexes a word or keyword and returns the corresponding token
     auto LexQuotedPhrase() -> Token;   // Lexes a quoted phrase and returns the corresponding token
     auto LexSymbol() -> Token;         // Lexes symbols (COLON, LPAREN, RPAREN) and returns the corresponding token
+    auto PeekWithoutConsuming() const -> std::vector<Token>; //Returns a vector of all tokens without consuming
 };
 
 #endif  // LEXER_H_