+++ CHANGELOG.md
@@ -72,6 +72,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Made GenericTermReader to union over all fields, not just search the body (#152)
 - add basic multiterm support for ranking (#153)
 - add scripts that made us build idx1(/5)
+- query coverage, density + position based ranking (#159)
 
 ### Fixed
 
@@ -111,11 +112,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - better integration of ranking, query & index (#137)
 - Fixed parser driver (#140)
 - ISRs and Position Index now use memory mapped files instead of reading byte-by-byte from ifstreams (#141)
+- Network support for query engine (#142)
 - Fixed parser drive (#143)
 - Changes from the runs on VM (perf/fixes for idx and ranking)(#132)
 - Fixed mmap bug in QueryEngine (#144)
 - Added favicon to front end (#150)
-- improve idx devops scripts
+- Added frequency count for queries (#151)
+- Added some code into QueryManager for token multiplicities (#152)
+- improve idx devops scripts (#153)
 
 ## [0.1.0] - 2025-02-07